Update Trajectory Create Utility - Added input units, weighted waypoints, and waypoint display
This is a preliminary update.  The Trajectory CSV file still has to have the comments modified to display the waypoint weights.  A lot of testing should also be done....  Otherwise this version has -- input unit selection, waypoint display, and optional weighted waypoints.  The functionality is now fairly identical to PathWeaver.  This util allows for all the constraints to be defined too...